fix(ci): restore the lint gate on main
Three lint steps failed on a pristine `main` checkout, so a required
status context could not distinguish a good pull request from a bad one
and every merge went through admin bypass.

- timer.rs had reached 2010 lines against the 2000-line cap. Its
  #[cfg(test)] scanner seeding/snapshot helpers move to
  timer/test_scanner_support.rs, re-exported by name because a glob does
  not propagate to the `crate::timer::...` call sites in the GC
  root-scanner tests. timer.rs is now 1866 lines.

- property_set.rs already had a GC_STORE_AUDIT(POINTER_FREE) marker for
  its guarded raw-f64 class-field store, but a multi-line
  canonicalize_raw_f64_numeric_store_value call had pushed it 8 lines
  above the store, outside the inventory's +/-6 window. The marker moves
  against the store it describes; nothing about the store changes.

- object/spill.rs (5 bare reads, ceiling 3) and json_tape.rs (22,
  ceiling 21) exceeded their raw-handle ceilings. Three reads convert to
  RuntimeHandle::across_mut rather than raising a ceiling, which is what
  raw_handle_debt_files.txt asks for: across object_meta_ensure and
  js_array_alloc_with_length_exact in spill.rs, and across the
  LazyArrayRooted safepoint in json_tape.rs. All three already reloaded
  the pointer afterwards, so this is the same discipline in the form the
  ratchet can count, and the pre-call address is no longer nameable.

Baseline drops 998 -> 993. --update also tightened three ceilings
carrying pre-existing slack (define_property.rs 3->2, reflect_support.rs
4->3, string/split.rs 10->7); no open pull request touches those files.

Refs #8092
