+++ crates/perry-runtime/src/object/spill.rs
@@ -180,16 +180,15 @@ pub(crate) fn reserve_object_spill(obj_ptr: usize, field_count: u32) {
 
         let scope = crate::gc::RuntimeHandleScope::new();
         let obj_handle = scope.root_raw_mut_ptr(obj);
-        object_meta_ensure(obj);
-
-        let obj = obj_handle.get_raw_mut_ptr::<ObjectHeader>();
+        let (_, obj) = obj_handle.across_mut::<ObjectHeader, _>(|| object_meta_ensure(obj));
         let meta = (*obj).meta;
         if (*meta).spill != 0 {
             return;
         }
 
-        let spill = crate::array::js_array_alloc_with_length_exact(field_count);
-        let obj = obj_handle.get_raw_mut_ptr::<ObjectHeader>();
+        let (spill, obj) = obj_handle.across_mut::<ObjectHeader, _>(|| {
+            crate::array::js_array_alloc_with_length_exact(field_count)
+        });
         let meta = (*obj).meta;
         if (*meta).spill == 0 {
             (*meta).spill = spill as u64;
